Pal 394 moved resource in example yaml, moved resource field from PolicyConfig class to PolicyCacheWarmer class, ammended return type for implementation of policyWarm method
