+++ example-library/src/main/java/uk/gov/gchq/palisade/example/config/ExamplePolicyCacheWarmerFactory.java
@@ -16,25 +16,38 @@
 
 package uk.gov.gchq.palisade.example.config;
 
+import org.apache.avro.reflect.MapEntry;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import uk.gov.gchq.palisade.Generated;
 import uk.gov.gchq.palisade.example.hrdatagenerator.types.Employee;
+import uk.gov.gchq.palisade.example.util.ExampleFileUtil;
+import uk.gov.gchq.palisade.resource.ParentResource;
 import uk.gov.gchq.palisade.resource.Resource;
+import uk.gov.gchq.palisade.resource.impl.DirectoryResource;
+import uk.gov.gchq.palisade.resource.impl.FileResource;
+import uk.gov.gchq.palisade.resource.impl.SystemResource;
 import uk.gov.gchq.palisade.rule.Rule;
 import uk.gov.gchq.palisade.service.PolicyCacheWarmerFactory;
 import uk.gov.gchq.palisade.service.UserCacheWarmerFactory;
 import uk.gov.gchq.palisade.service.request.Policy;
 
 import java.lang.reflect.InvocationTargetException;
+import java.net.URI;
+import java.nio.file.FileSystems;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.Map.Entry;
 import java.util.Objects;
 import java.util.StringJoiner;
+import java.util.stream.StreamSupport;
 
+import static java.util.Objects.isNull;
 import static java.util.Objects.requireNonNull;
 
 @ConfigurationProperties
@@ -43,6 +56,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     private static final Logger LOGGER = LoggerFactory.getLogger(ExamplePolicyCacheWarmerFactory.class);
 
     private String type;
+    private String resource;
     private String owner;
     private Map<String, String> resourceRules;
     private Map<String, String> recordRules;
@@ -62,13 +76,16 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
      * Constructor with 4 arguments for an example implementation
      * of the {@link PolicyCacheWarmerFactory} interface
      *
-     * @param type          a {@link String} of the {@link Policy} type.
+     * @param type          a {@link String} value of the {@link Policy} type.
+     * @param resource      a {@link String} value of the {@link Resource} to be used.
      * @param owner         a {@link String} value of the owner of the policy
      * @param resourceRules a {@link Map} containing the ({@link String}) message and the ({@link String}) rule name.
      * @param recordRules   a {@link Map} containing the ({@link String}) message and the ({@link String}) rule name.
      */
-    public ExamplePolicyCacheWarmerFactory(final String type, final String owner, final Map<String, String> resourceRules, final Map<String, String> recordRules) {
+    public ExamplePolicyCacheWarmerFactory(final String type, final String resource, final String owner,
+                                           final Map<String, String> resourceRules, final Map<String, String> recordRules) {
         this.type = type;
+        this.resource = resource;
         this.owner = owner;
         this.resourceRules = resourceRules;
         this.recordRules = recordRules;
@@ -83,6 +100,17 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     public void setType(final String type) {
         requireNonNull(type);
         this.type = type;
+    }
+
+    @Generated
+    public String getResource() {
+        return resource;
+    }
+
+    @Generated
+    public void setResource(final String resource) {
+        requireNonNull(resource);
+        this.resource = resource;
     }
 
     @Generated
@@ -119,7 +147,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     }
 
     @Override
-    public Policy<Employee> policyWarm(final List<? extends UserCacheWarmerFactory> users) {
+    public Entry<Resource, Policy> policyWarm(final List<? extends UserCacheWarmerFactory> users) {
         Policy<Employee> policy = new Policy<>();
         for (ExampleUserCacheWarmerFactory user : (List<ExampleUserCacheWarmerFactory>) users) {
             if (user.getUserId().equals(owner)) {
@@ -132,13 +160,13 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
         for (String key : recordRules.keySet()) {
             policy.recordLevelRule(key, (Rule<Employee>) createRule(recordRules.get(key), "record"));
         }
-        return policy;
+        return new MapEntry<>(createResource(), policy);
     }
 
     private Rule<?> createRule(final String rule, final String ruleType) {
         if ("resource".equalsIgnoreCase(ruleType)) {
             try {
-                LOGGER.debug("{} - {}", rule, ruleType);
+                LOGGER.debug("Adding rule {} for rule type {}", rule, ruleType);
                 return (Rule<Resource>) Class.forName(rule).getConstructor().newInstance();
             } catch (ClassNotFoundException | NoSuchMethodException ex) {
                 LOGGER.error("Error getting class: {}", ex.getMessage());
@@ -152,7 +180,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
         }
         if ("record".equalsIgnoreCase(ruleType)) {
             try {
-                LOGGER.debug("{} - {}", rule, ruleType);
+                LOGGER.debug("Adding rule {} for rule type {}", rule, ruleType);
                 return (Rule<Employee>) Class.forName(rule).getConstructor().newInstance();
             } catch (ClassNotFoundException | NoSuchMethodException ex) {
                 LOGGER.error("Error getting class: {}", ex.getMessage());
@@ -168,6 +196,47 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     }
 
     @Override
+    public Resource createResource() {
+        URI normalised = ExampleFileUtil.convertToFileURI(resource);
+        String resource = normalised.toString();
+        if (resource.endsWith(".avro")) {
+            return new FileResource().id(resource).type(Employee.class.getTypeName()).serialisedFormat("avro").parent(getParent(resource));
+        } else {
+            return new DirectoryResource().id(resource).parent(getParent(resource));
+        }
+    }
+
+    public ParentResource getParent(final String fileURL) {
+        URI normalised = ExampleFileUtil.convertToFileURI(fileURL);
+        //this should only be applied to URLs that start with 'file://' not other types of URL
+        if (normalised.getScheme().equals(FileSystems.getDefault().provider().getScheme())) {
+            Path current = Paths.get(normalised);
+            Path parent = current.getParent();
+            //no parent can be found, must already be a directory tree root
+            if (isNull(parent)) {
+                throw new IllegalArgumentException(fileURL + " is already a directory tree root");
+            } else if (isDirectoryRoot(parent)) {
+                //else if this is a directory tree root
+                return new SystemResource().id(parent.toUri().toString());
+            } else {
+                //else recurse up a level
+                return new DirectoryResource().id(parent.toUri().toString()).parent(getParent(parent.toUri().toString()));
+            }
+        } else {
+            //if this is another scheme then there is no definable parent
+            return new SystemResource().id("");
+        }
+    }
+
+    public boolean isDirectoryRoot(final Path path) {
+        return StreamSupport
+                .stream(FileSystems.getDefault()
+                        .getRootDirectories()
+                        .spliterator(), false)
+                .anyMatch(path::equals);
+    }
+
+    @Override
     @Generated
     public boolean equals(final Object o) {
         if (this == o) {
@@ -178,6 +247,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
         }
         final ExamplePolicyCacheWarmerFactory that = (ExamplePolicyCacheWarmerFactory) o;
         return Objects.equals(type, that.type) &&
+                Objects.equals(resource, that.resource) &&
                 Objects.equals(owner, that.owner) &&
                 Objects.equals(resourceRules, that.resourceRules) &&
                 Objects.equals(recordRules, that.recordRules);
@@ -186,7 +256,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     @Override
     @Generated
     public int hashCode() {
-        return Objects.hash(type, owner, resourceRules, recordRules);
+        return Objects.hash(type, resource, owner, resourceRules, recordRules);
     }
 
     @Override
@@ -194,6 +264,7 @@ public class ExamplePolicyCacheWarmerFactory implements PolicyCacheWarmerFactory
     public String toString() {
         return new StringJoiner(", ", ExamplePolicyCacheWarmerFactory.class.getSimpleName() + "[", "]")
                 .add("type='" + type + "'")
+                .add("resource='" + resource + "'")
                 .add("owner='" + owner + "'")
                 .add("resourceRules=" + resourceRules)
                 .add("recordRules=" + recordRules)
