+++ example-library/src/main/java/uk/gov/gchq/palisade/example/config/ExamplePolicyConfiguration.java
@@ -19,26 +19,15 @@ package uk.gov.gchq.palisade.example.config;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import uk.gov.gchq.palisade.Generated;
-import uk.gov.gchq.palisade.example.hrdatagenerator.types.Employee;
 import uk.gov.gchq.palisade.example.util.ExampleFileUtil;
-import uk.gov.gchq.palisade.resource.ParentResource;
-import uk.gov.gchq.palisade.resource.Resource;
-import uk.gov.gchq.palisade.resource.impl.DirectoryResource;
-import uk.gov.gchq.palisade.resource.impl.FileResource;
-import uk.gov.gchq.palisade.resource.impl.SystemResource;
 import uk.gov.gchq.palisade.service.PolicyConfiguration;
 
 import java.net.URI;
-import java.nio.file.FileSystems;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
 import java.util.StringJoiner;
-import java.util.stream.StreamSupport;
 
-import static java.util.Objects.isNull;
 import static java.util.Objects.requireNonNull;
 
 @ConfigurationProperties(prefix = "population")
@@ -106,47 +95,6 @@ public class ExamplePolicyConfiguration implements PolicyConfiguration {
     public void setUsers(final List<ExampleUserCacheWarmerFactory> users) {
         requireNonNull(users);
         this.users = users;
-    }
-
-    @Override
-    public Resource createResource() {
-        URI normalised = ExampleFileUtil.convertToFileURI(resource);
-        String resource = normalised.toString();
-        if (resource.endsWith(".avro")) {
-            return new FileResource().id(resource).type(Employee.class.getTypeName()).serialisedFormat("avro").parent(getParent(resource));
-        } else {
-            return new DirectoryResource().id(resource).parent(getParent(resource));
-        }
-    }
-
-    public ParentResource getParent(final String fileURL) {
-        URI normalised = ExampleFileUtil.convertToFileURI(fileURL);
-        //this should only be applied to URLs that start with 'file://' not other types of URL
-        if (normalised.getScheme().equals(FileSystems.getDefault().provider().getScheme())) {
-            Path current = Paths.get(normalised);
-            Path parent = current.getParent();
-            //no parent can be found, must already be a directory tree root
-            if (isNull(parent)) {
-                throw new IllegalArgumentException(fileURL + " is already a directory tree root");
-            } else if (isDirectoryRoot(parent)) {
-                //else if this is a directory tree root
-                return new SystemResource().id(parent.toUri().toString());
-            } else {
-                //else recurse up a level
-                return new DirectoryResource().id(parent.toUri().toString()).parent(getParent(parent.toUri().toString()));
-            }
-        } else {
-            //if this is another scheme then there is no definable parent
-            return new SystemResource().id("");
-        }
-    }
-
-    public boolean isDirectoryRoot(final Path path) {
-        return StreamSupport
-                .stream(FileSystems.getDefault()
-                        .getRootDirectories()
-                        .spliterator(), false)
-                .anyMatch(path::equals);
     }
 
     private String createFilePath(final String file) {
