+++ example-library/src/main/resources/application-example.yaml
@@ -33,9 +33,9 @@ population:
         public
       roles:
         IT
-  resource: ../Palisade-examples/resources/data
   policies:
     - type: Employee
+      resource: ../Palisade-examples/resources/data
       owner: Alice
       recordRules:
         "1-Bank details only visible for the purpose of salary": uk.gov.gchq.palisade.example.rule.BankDetailsRule
